requirements.txt
autogen
transformers
torch
faiss-cpu
sentence-transformers
langchain
python-dotenv
langchain-community
fastapi 
uvicorn 
joblib